% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerfDiNet.R
\name{calcPerfDiNet}
\alias{calcPerfDiNet}
\title{Calculating performance metrics of the directed net 'inferredNet' w.r.t. the
directed net 'targetNet'.}
\usage{
calcPerfDiNet(inferredNet, targetNet, Result, n)
}
\arguments{
\item{inferredNet}{directed net 'inferredNet'}

\item{targetNet}{directed net 'targetNet'}

\item{Result}{Matrix to store the results}

\item{n}{Number of nodes in each of the nets.}
}
\value{
the performance metrics
}
\description{
Calculating performance metrics of the directed net 'inferredNet' w.r.t. the
directed net 'targetNet'.
}
\examples{
res<-calcPerfDiNet(matrix(c(1,0,1,1,1,1,1,0,1),nrow=3,ncol=3),
+ matrix(c(0,1,0,1,1,1,1,0,1),nrow=3,ncol=3),
+ matrix(nrow=1,ncol=11),3)

}
