% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUnrolledDbn.R
\name{checkUnrolledDbn}
\alias{checkUnrolledDbn}
\title{Checks whether the given unrolled DBN follows 1st Markov order or not}
\usage{
checkUnrolledDbn(unrolled.DBN.adj.matrix)
}
\arguments{
\item{unrolled.DBN.adj.matrix}{An unrolled DBN adjacency matrix}
}
\value{
whether the given unrolled DBN follows 1st Markov order or not
}
\description{
Checks whether the given unrolled DBN follows 1st Markov order or not
}
\examples{
checkUnrolledDbn(matrix(c(0,0,0,0),nrow=2,ncol=2))

}
