% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareNet.R
\name{CompareNet}
\alias{CompareNet}
\title{Checks if 'di.net.adj.matrix' = 'cmi.net.adj.matrix'}
\usage{
CompareNet(di.net.adj.matrix, cmi.net.adj.matrix, num.node)
}
\arguments{
\item{di.net.adj.matrix}{First adjacency metric}

\item{cmi.net.adj.matrix}{Second adjacency metric}

\item{num.node}{Number of nodes in the metrices}
}
\value{
Returns 1 if 'di.net.adj.matrix' = 'cmi.net.adj.matrix' else returns 0
}
\description{
Checks if 'di.net.adj.matrix' = 'cmi.net.adj.matrix'
}
\examples{
CompareNet(matrix(c(0,0,0,0),nrow=2,ncol=2),
+ matrix(c(0,0,0,0),nrow=2,ncol=2),
+ 2)

CompareNet(matrix(c(0,0,1,0),nrow=2,ncol=2),
+ matrix(c(0,0,0,0),nrow=2,ncol=2),
+ 2)
}
