% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondMutualInfo.R
\name{computeCmi}
\alias{computeCmi}
\title{Compute Conditional Mutual Infortion (CMI)}
\usage{
computeCmi(v1, v2, vcs)
}
\arguments{
\item{v1}{random variable 1}

\item{v2}{random variable 2}

\item{vcs}{random vector 'vcs'}
}
\value{
Conditional Mutual info between 'v1' and 'v2'
}
\description{
Compute Conditional Mutual Info between random variables 'v1' and 'v2'
given random vector 'vcs'.
}
\examples{
computeCmi(c(3,5),c(3,5))

}
