% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{ComputEntropy}
\alias{ComputEntropy}
\title{Compute Entropy matrix from the input data}
\usage{
ComputEntropy(input.data)
}
\arguments{
\item{input.data}{input data}
}
\value{
entropy matrix
}
\description{
Compute Entropy matrix from the input data
}
