% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dinet_to_undinet.R
\name{ConvertDinetToUndinet}
\alias{ConvertDinetToUndinet}
\title{Given a directed network, convert it into an undirected network}
\usage{
ConvertDinetToUndinet(di.net)
}
\arguments{
\item{di.net}{a directed network}
}
\value{
directed network converted to undirected network
}
\description{
Given a directed network, convert it into an undirected network
}
\examples{
ConvertDinetToUndinet(matrix(c(0,1,0,0),nrow=2))

}
