% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_feed_fwd_edges_undi.R
\name{CountFeedFwdEdgesUndi}
\alias{CountFeedFwdEdgesUndi}
\title{Count the number of feed-forward edges in a given undirected network.}
\usage{
CountFeedFwdEdgesUndi(undi.net.adj.matrix)
}
\arguments{
\item{undi.net.adj.matrix}{adjacency matrix of an undirected network}
}
\value{
count of the number of feed-forward edges
}
\description{
Count the number of feed-forward edges in a given undirected network.
}
\examples{
CountFeedFwdEdgesUndi(matrix(c(0,1,0,1,0,1,0,0,0),nrow=3))

}
