% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeData.R
\name{discretizeData.2L.Tesla}
\alias{discretizeData.2L.Tesla}
\title{Discretize input data into 2 levels.}
\usage{
discretizeData.2L.Tesla(input.data)
}
\arguments{
\item{input.data}{data frame to be discretized. rows have samples and cols have variables.}
}
\value{
data discretized into 2 levels
}
\description{
Discretize input data into the following two levels as done in
For each gene, the expression values are first sorted; then the top 2
extreme values in either end of the sorted list are discarded; last,
the median of the remaining values is used as the threshold above
which the value is binarized as 'Level 2' and 'Level 1' otherwise. Here, 2 means
the expression of a gene is up-regulated, and 1 means down-regulated.
}
\references{
1. Ahmed, Amr, and Eric P. Xing. "Recovering time-varying networks of dependencies in social and biological studies."
Proceedings of the National Academy of Sciences 106.29 (2009): 11878-11883.]
}
