% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeData.R
\name{discretizeData.2L.wt.le}
\alias{discretizeData.2L.wt.le}
\title{Discretizes input data into two levels.}
\usage{
discretizeData.2L.wt.le(input.data, input.wt.data.filename)
}
\arguments{
\item{input.data}{the data that is given as the input, data frame to be discretized}

\item{input.wt.data.filename}{path of the file containing the table}
}
\value{
input data discretized into 2 levels
}
\description{
Discretizes input data into the following two levels:
Less than wild type => level 1
Greater than equal to wild type => level 2
}
