% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerfDiNet.R
\name{eval.wrt.known.gene.ias}
\alias{eval.wrt.known.gene.ias}
\title{Accuracy of predicted directed gene reuglatory network adjacency matrix}
\usage{
eval.wrt.known.gene.ias(di.net.adj.matrix)
}
\arguments{
\item{di.net.adj.matrix}{predicted directed gene reuglatory network adjacency matrix}
}
\description{
Given a predicted directed gene reuglatory network adjacency matrix, evaluate its
accuracy w.r.t. known gene-gene interactions.
}
\references{
[1] Song, Le, Mladen Kolar, and Eric P. Xing. "KELLER: estimating time-varying interactions
between genes." Bioinformatics 25.12 (2009): i128-i136.
}
