% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTrueNetFunc.R
\name{GenTrueAdjMatrix}
\alias{GenTrueAdjMatrix}
\title{Generates True net adjacency matrix and save as an R object}
\usage{
GenTrueAdjMatrix(input.file, output.file, num.nodes)
}
\arguments{
\item{input.file}{full path of the input file containing adjacency list}

\item{output.file}{full path where output adjacency matrix is to be stored}

\item{num.nodes}{the number of nodes}
}
\value{
also returns the adjacency matrix
}
\description{
Generates True net adjacency matrix and save as an R object
}
