% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{LearnClr3NetMfi}
\alias{LearnClr3NetMfi}
\title{Learn CLR3 network}
\usage{
LearnClr3NetMfi(input.data.discr.3D, num.nodes, node.names, num.timepts,
  max.fanin, mi.net.adj.matrix.list)
}
\arguments{
\item{input.data.discr.3D}{3D Discretized dataset}

\item{num.nodes}{number of nodes}

\item{node.names}{name of nodes}

\item{num.timepts}{number of timepoints}

\item{max.fanin}{number of top neighbours to retain}

\item{mi.net.adj.matrix.list}{mi network adjacency matrix list}
}
\value{
mi network adjacency matrix list
}
\description{
Learns CLR3 net. For each node, retains top 'max.fanin' number of neighbours w.r.t. edge weight
and removes rest of the edges. Tie is broken in favour of the neighbour having smaller node index.
If there are less than that number of edges for a node, then retain all its neighbours.
}
