% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{LearnClrNetMfi}
\alias{LearnClrNetMfi}
\title{Learns CLR network}
\usage{
LearnClrNetMfi(mut.info.matrix, mi.net.adj.matrix, num.nodes, max.fanin,
  output.dirname = "./OUTPUT")
}
\arguments{
\item{mut.info.matrix}{matrix containing mut info}

\item{mi.net.adj.matrix}{mi network adjacency matrix}

\item{num.nodes}{number of nodes}

\item{max.fanin}{number of top neighbours to retain}

\item{output.dirname}{output directory to store files}
}
\value{
mi network adjacency matrix
}
\description{
Learns CLR net. For each node, retains top 'max.fanin' number of neighbours w.r.t. edge weight
and removes rest of the edges. Tie is broken in favour of the neighbour having smaller node index.
If there are less than that number of edges for a node, then retain all its neighbours.
}
