% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnCmiNetStruct.R
\name{learnCmiNetStruct}
\alias{learnCmiNetStruct}
\title{Learns the CMI structure}
\usage{
learnCmiNetStruct(di.net.adj.matrix, input.data, num.nodes, beta)
}
\arguments{
\item{di.net.adj.matrix}{a directed network adjacency matrix}

\item{input.data}{data given as input data frame}

\item{num.nodes}{the number of nodes}

\item{beta}{value of parameter beta}
}
\value{
the directed network adjacency matrix after learning the CMI network
}
\description{
Learns the CMI structure
}
