% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnDbnStruct3dParDeg1.R
\name{LearnDbnStructMo1Clr3Ser}
\alias{LearnDbnStructMo1Clr3Ser}
\title{Learns DBN structure of Markov order 1 where candidate parents are selected using the CLR3 algo.}
\usage{
LearnDbnStructMo1Clr3Ser(input.data.discr.3D,
  mi.net.adj.matrix.list.filename, num.discr.levels, num.nodes,
  num.timepts, max.fanin, node.names, unrolled.DBN.adj.matrix.list)
}
\arguments{
\item{input.data.discr.3D}{Dimensions {1 = time points, 2 = variables, 3 = samples under the same time point}.}

\item{mi.net.adj.matrix.list.filename}{List of filenames (relative/absolute path) of adjacency matrices of the mutual information network.}

\item{num.discr.levels}{If input data is discretized, then number of discrete levels for each variable. Else if input data is continuous, then number of levels in which data needs to be discretized for performing the DBN structure learning.}

\item{num.nodes}{number of nodes}

\item{num.timepts}{number of timepoints}

\item{max.fanin}{maximum incoming edges in the graph}

\item{node.names}{name of the nodes}

\item{unrolled.DBN.adj.matrix.list}{List of unrolled DBN adjacency matrices}
}
\value{
Unrolled DBN adjacency matrix
}
\description{
Learns DBN structure of Markov order 1 where candidate parents are selected using the CLR3 algo.
It is a serial algorithmic implementation.
}
