% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{LearnMiNetStructClr}
\alias{LearnMiNetStructClr}
\title{Learns the CLR network Replaces all non-zero edge weights with 1.}
\usage{
LearnMiNetStructClr(mut.info.matrix, mi.net.adj.matrix, num.nodes,
  output.dirname = "./OUTPUT")
}
\arguments{
\item{mut.info.matrix}{matrix containing mut info}

\item{mi.net.adj.matrix}{mi network adjacency matrix}

\item{num.nodes}{number of nodes}

\item{output.dirname}{output directory to store files}
}
\value{
mi network adjacency matrix
}
\description{
Learns the CLR network Replaces all non-zero edge weights with 1.
}
