% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{LearnMiNetStructRowMedian}
\alias{LearnMiNetStructRowMedian}
\title{Learn the mi network structure}
\usage{
LearnMiNetStructRowMedian(mut.info.matrix, mi.net.adj.matrix, num.nodes)
}
\arguments{
\item{mut.info.matrix}{matrix containing mut info}

\item{mi.net.adj.matrix}{mi network adjacency matrix}

\item{num.nodes}{number of nodes}
}
\value{
mi network adjacency matrix
}
\description{
Learn the mi network structure
}
