% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{LearnMiNetStructZstat}
\alias{LearnMiNetStructZstat}
\title{Learn the mi network structure}
\usage{
LearnMiNetStructZstat(mut.info.matrix, mi.net.adj.matrix, entropy.matrix,
  alpha)
}
\arguments{
\item{mut.info.matrix}{matrix containing mut info}

\item{mi.net.adj.matrix}{mi network adjacency matrix}

\item{entropy.matrix}{matrix containing the entropy information}

\item{alpha}{parameter value alpha}
}
\value{
mi network adjacency matrix
}
\description{
Learn the mi network structure
}
