% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareNet.R
\name{Print.common.di.edges}
\alias{Print.common.di.edges}
\title{Given two di network adjacency matrices, it prints the common edges in an output file}
\usage{
Print.common.di.edges(di.net.adj.matrix1, di.net.adj.matrix2,
  output.dirname = "./OUTPUT")
}
\arguments{
\item{di.net.adj.matrix1}{first di network adjacency matrix}

\item{di.net.adj.matrix2}{second di network adjacency matrix}

\item{output.dirname}{output directory to store files}
}
\description{
Given two di nets' adjacency matrices (must of of same dim, same rownames, same colnames
where rows = source nodes, cols = tgt nodes, 0 and 1 represent absence and presence of
an edge, resp.), this function prints the common edges in an output text file.
}
