% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGS-package.R
\docType{package}
\name{TGS-package}
\alias{TGS-package}
\alias{TGS}
\title{TGS: A package for Rapid Reconstruction of Time-Varying Gene Regulatory
Networks}
\description{
The \code{TGS} package provides an implementation of the TGS algorithm
and its variants. This algorithm reconstructs time-varying gene
regulatory networks from time-series gene expression datasets. For
algorithmic details, please see: Saptarshi Pyne, Alok Ranjan Kumar,
and Ashish Anand. Rapid reconstruction of time-varying gene
regulatory networks. IEEE/ACM Transactions on Computational Biology
and Bioinformatics, 17(1):278--291, Jan--Feb 2020.
}
\section{The \code{LearnTgs} function}{

Please call \code{TGS::LearnTgs()} to invoke the TGS algorithm. See
the manual for the required parameters.
}

