% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGS.R
\docType{package}
\name{TGS}
\alias{TGS}
\title{TGS: A package for Rapid Reconstruction of Time-Varying Gene Regulatory
Networks}
\description{
The \code{TGS} package provides an implementation of the TGS algorithm (Pyne et
al., 2018, \url{doi:10.1109/TCBB.2018.2861698}). This algorithm reconstructs
time-varying gene regulatory networks from time-series gene expression
datasets.
}
\section{The \code{LearnTgs} function}{

Please call \code{TGS::LearnTgs()} to invoke the TGS algorithm. See
the manual for the required parameters.
}

