% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{ComputEntropy}
\alias{ComputEntropy}
\title{Compute Entropy matrix from the input data}
\usage{
ComputEntropy(input.data)
}
\arguments{
\item{input.data}{input data matrix}
}
\value{
entropy matrix
}
\description{
Compute Entropy matrix from the input data
}
\examples{
df = data.frame(c(2,3,5), c(1,3,2), c(2,31,4))
ComputEntropy(df)
}
