% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_mi_net_struct.R
\name{LearnMiNetStructRowMedian}
\alias{LearnMiNetStructRowMedian}
\title{Learn the mi network structure}
\usage{
LearnMiNetStructRowMedian(mut.info.matrix, mi.net.adj.matrix, num.nodes)
}
\arguments{
\item{mut.info.matrix}{matrix containing mut info}

\item{mi.net.adj.matrix}{mi network adjacency matrix}

\item{num.nodes}{number of nodes}
}
\value{
mi network adjacency matrix
}
\description{
Learn the mi network structure
}
\examples{
LearnMiNetStructRowMedian(
+ matrix(c(0.1,0.5,0.53,0.76,0,0.12,0.43,0.65,0.23),nrow=3),
+ matrix(c(1,0,1,0,0,0,1,0,1),nrow=3),
+ 3)

}
